+++ src/components/NavbarLeft/docs/NavbarLeft.test.tsx
@@ -1,9 +1,9 @@
-import React from "react";
-import { render } from "utils/quotly-testing";
-import NavbarLeft from "../NavbarLeft";
+import React from 'react';
+import { render } from 'utils/quotly-testing';
+import NavbarLeft from '../NavbarLeft';
 
-describe("NavbarLeft", () => {
-  it("renders", () => {
+describe('NavbarLeft', () => {
+  it('renders', () => {
     expect(render(<NavbarLeft />)).not.toBeNull();
   });
 });