+++ src/setupTests.ts
@@ -7,3 +7,7 @@ import '__mocks__/window';
 import '__mocks__/iconify-react';
 
 jest.mock('react-markdown');
+jest.mock('react-i18next', () => ({
+  useTranslation: () => ({ t: (key: string) => key }),
+}));
+jest.mock('@fortawesome/react-fontawesome');