+++ changelog.md
@@ -2,12 +2,16 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.2.1] - 2024-10-12
+### Additions
+- Added i18n for translations [#4](https://github.com/quotly-eu/quotly-frontend/issues/4)
+
 ### Changes
 - theme file now has get values for font sizes
 - Switched from manual PWA to Vite-plugin-pwa
 - Button now vibrate shortly on click
 - Disabled ContextMenu
 - dated now uses date for locale formatting
+- 404 page has now been updated
 
 ### Fixes
 - Adjusted mobile responsiveness